Adapt to linked angular CLI
