--- frontend/app/components/projects-overview/overview-block.directive.ts
@@ -1,75 +1,0 @@
-// -- copyright
-// OpenProject is a project management system.
-// Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
-//
-// This program is free software; you can redistribute it and/or
-// modify it under the terms of the GNU General Public License version 3.
-//
-// OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
-// Copyright (C) 2006-2013 Jean-Philippe Lang
-// Copyright (C) 2010-2013 the ChiliProject Team
-//
-// This program is free software; you can redistribute it and/or
-// modify it under the terms of the GNU General Public License
-// as published by the Free Software Foundation; either version 2
-// of the License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program; if not, write to the Free Software
-// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-//
-// See doc/COPYRIGHT.rdoc for more details.
-// ++
-
-import {ProjectsOverviewController} from './overview-page-layout.directive';
-export class OverviewBlockController {
-
-  public layoutCtrl:ProjectsOverviewController;
-
-  constructor(public $element:ng.IAugmentedJQuery) {
-  }
-
-  public get editing() {
-    return !!this.layoutCtrl;
-  }
-
-  /**
-   * Remove this block
-   */
-  public remove() {
-    this.$element.remove();
-  }
-
-}
-
-function overviewBlock():any {
-  return {
-    restrict: 'EA',
-    scope: {},
-    transclude: true,
-    compile: function() {
-      return function(
-        scope:any,
-        element:ng.IAugmentedJQuery,
-        attrs:ng.IAttributes,
-        ctrl:any,
-        transclude:any) {
-        scope.$ctrl.layoutCtrl = ctrl;
-        transclude(scope, (clone:any) => {
-          element.append(clone);
-        });
-      };
-    },
-    require: '?^overviewPageLayout',
-    controller: OverviewBlockController,
-    bindToController: true,
-    controllerAs: '$ctrl'
-  };
-}
-
-angular.module('openproject').directive('overviewBlock', overviewBlock);
