+++ frontend/legacy-app/components/projects-overview/overview-textile-block.directive.ts
@@ -27,6 +27,8 @@
 // ++
 
 import {ProjectsOverviewController} from './overview-page-layout.directive';
+import {PluginContextService} from "core-app/services/plugin-context.service";
+
 export class OverviewTextileBlockController {
 
   public layoutCtrl:ProjectsOverviewController;
@@ -47,8 +49,8 @@ export class OverviewTextileBlockController {
               public $timeout:ng.ITimeoutService,
               public $window:ng.IWindowService,
               public $q:ng.IQService,
-              public I18n:op.I18n,
-              public NotificationsService:any) {
+              public I18n:any,
+              public pluginContext:PluginContextService) {
   }
 
   public initialize() {
@@ -99,14 +101,14 @@ export class OverviewTextileBlockController {
     }).done((response) => {
       deferred.resolve();
       this.$timeout(() => {
-        this.NotificationsService.addSuccess(this.I18n.t('js.notice_successful_update'));
+        this.pluginContext.context!.services.notifications.addSuccess(this.I18n.t('js.notice_successful_update'));
       });
       this.layoutCtrl.updateBlock(this.blockName, response);
       this.layoutCtrl.updateAttachments();
     }).fail((error) => {
       deferred.reject();
       this.$timeout(() => {
-        this.NotificationsService.addError(
+        this.pluginContext.context!.services.addError(
           this.I18n.t('js.notification_update_block_failed') + ' ' + error.responseText
         );
       });
@@ -146,4 +148,4 @@ function overviewTextileBlock():any {
   };
 }
 
-angular.module('openproject').directive('overviewTextileBlock', overviewTextileBlock);
+angular.module('OpenProjectLegacy').directive('overviewTextileBlock', overviewTextileBlock);
