+++ spec/features/project_overview_spec.rb
@@ -26,7 +26,7 @@
 # See doc/COPYRIGHT.rdoc for more details.
 #++
 
-require File.dirname(__FILE__) + '/../spec_helper'
+require 'spec_helper'
 
 describe 'My project page overview', type: :feature do
   let(:project) { FactoryBot.create :project }
